Login.logout, session tracking implemented

+++ java-web-feb/src/com/assimilate/javaweb/servlet/ViewAllEmployeesServlet.java
@@ -7,11 +7,13 @@ import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
 
+import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 /**
  * Servlet implementation class ViewAllEmployeesServlet
@@ -33,35 +35,52 @@ public class ViewAllEmployeesServlet extends HttpServlet {
 	 */
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		
-		Connection connection = JdbcTest.getConnection();
-		String query = "select * from employee";
-		try {
-			Statement statement = connection.createStatement();
-			ResultSet resultSet = statement.executeQuery(query);
-			response.setContentType("text/html");
-			PrintWriter printWriter = response.getWriter();
-			printWriter.write("<table \"width:50%, border=1\">");
-			printWriter.write("<tr>\n" + 
-					"    <th>Id</th>\n" + 
-					"    <th>name</th>\n" + 
-					"    <th>Age</th>\n" + 
-					"    <th>Action</th>\n" + 
-					"  </tr>");
-			while (resultSet.next()) {
+		response.setContentType("text/html");
+		PrintWriter printWriter = response.getWriter();
+		HttpSession session = request.getSession();
+		String userFromSession = (String) session.getAttribute("loggedInUser");
+		if(userFromSession == null) {
+			
+			printWriter.write("Please login first.");
+			RequestDispatcher requestDispatcher = request.getRequestDispatcher("login.html");
+			requestDispatcher.include(request, response);
+			return;
+		}else {
+
+			Connection connection = JdbcTest.getConnection();
+			String query = "select * from employee";
+			try {
+				Statement statement = connection.createStatement();
+				ResultSet resultSet = statement.executeQuery(query);
+				RequestDispatcher requestDispatcher = request.getRequestDispatcher("welcome.html");
+				requestDispatcher.include(request, response);
+				
+				printWriter.write("<table \"width:50%, border=1\">");
 				printWriter.write("<tr>\n" + 
-						"    <td>"+resultSet.getInt("id")+"</td>\n" + 
-						"    <td>"+resultSet.getString("name")+"</td>\n" + 
-						"    <td>"+ resultSet.getInt("age") + "</td>\n" +
-						"    <td> <a href='DeleteEmployeeServlet?id="+resultSet.getInt("id")+"'>Delete </a> </td>\n" +
-						"    <td> <a href='EditEmployeeServlet?id="+resultSet.getInt("id")+"'>Edit</a> </td>\n"
-						+ "  </tr>");
+						"    <th>Id</th>\n" + 
+						"    <th>name</th>\n" + 
+						"    <th>Age</th>\n" + 
+						"    <th>Action</th>\n" + 
+						"  </tr>");
+				while (resultSet.next()) {
+					printWriter.write("<tr>\n" + 
+							"    <td>"+resultSet.getInt("id")+"</td>\n" + 
+							"    <td>"+resultSet.getString("name")+"</td>\n" + 
+							"    <td>"+ resultSet.getInt("age") + "</td>\n" +
+							"    <td> <a href='DeleteEmployeeServlet?id="+resultSet.getInt("id")+"'>Delete </a> </td>\n" +
+							"    <td> <a href='EditEmployeeServlet?id="+resultSet.getInt("id")+"'>Edit</a> </td>\n"
+							+ "  </tr>");
+				}
+				printWriter.write("</table>");
+
+			} catch (SQLException e) {
+				e.printStackTrace();
 			}
-			printWriter.write("</table>");
 
-		} catch (SQLException e) {
-			e.printStackTrace();
 		}
-
+		
+		
+		
 	}
 
 	/**
